+++ app/src/androidTest/java/com/android/sample/screen/AddStudyTaskModalTest.kt
@@ -1,0 +1,3 @@
+package com.android.sample.screen
+
+class AddStudyTaskModalTest {}
